+++ Backend/src/controllers/homePageController.ts
